requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
markovify==0.8.0
multidict==4.7.6
Unidecode==1.1.1
websockets==8.1
yarl==1.4.2
